pyproject.toml
[tool.poetry]
name = "parafia-szopka"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<=3.12"
python-kasa = "^0.5.4"
python-vlc = "^3.0.20123"
pyserial = "^3.5"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.27.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
